The answer are now submitted and evaluated
Changed the Answering Service so that the submited answer are evaluated
with eval(question). Previously, questions from the server can be any
free text. Now they must be in the form that is a math expression that
can be evaluated. eg. '6 + 4', or '10 x 2'. The answering service will
replace the 'x' and '÷' with '*' and '/'. So they can be evaluated.