Provide Feedback for answer
Green background that flashes for correct answer. Red permenent
background for incorrect answer. Because the player need to answer the
question again. Also added message at the bottom of the input field that
says 'correct' or 'ops, try again'. The progress bar has also been
updated to show who is winning. Worst case senario, the progress bar
will show 98% to 2% so both player's score are still showing on the bar.