Built in an AI
Reformated how variables are stored so they are all inside an object
called p. so instead of myAnswer and theirAnswer. now it's all
p.User.answer and p.Opponent.answer

Changed how the scoring system works. Now that when the sum of both
players points are 15, then whenever one person scores, the other person
looses a point. This is to make the progress bar look more dynamic and
intense.

Built an AI that answers questions at random intervals