Timer and Game States
Three game states have been added and they are controlled by two
booleans GameStarted and GameEnded.

when game starts, timer starts. when game ends, a score board shows up.

When score board is closed, all the stats are reset thus goin into the
pre-starting position.

The timer is added. and the bot has been adjusted so that if the player
is losing, the bot will go slightly easier on him/her.